testing: use an adapting layer to be comaptible with multiple versions of PHP
