+++ tests/Common/MultiPHPUnitVersionAdapter_5_7.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace DDTrace\Tests\Common;
+
+use PHPUnit\Framework\TestCase;
+
+abstract class MultiPHPUnitVersionAdapter extends TestCase
+{
+    abstract protected function afterSetUp();
+
+    abstract protected function beforeTearDown();
+
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->afterSetUp();
+    }
+
+    protected function tearDown()
+    {
+        $this->beforeTearDown();
+        parent::tearDown();
+    }
+}
